moved _find_outliers in the new vision_utilities.py file.
